Lesson-3: Интерфейсы репозиториев вынесены из файлов репозиториев
